+++ src/Modules/Prompt.react.js
@@ -3,15 +3,21 @@
 import _ from 'lodash';
 import ClassNames from 'classnames';
 import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
 
 class Prompt extends Component {
 
     constructor(props) {
         super(props);
 
-        this.state = { show: props.show || false };
+        this.state = {
+            show: props.show || false,
+            inlineVerticalAlign: 0
+        };
 
         this._onClick = this._onClick.bind(this);
+        this._onNoClick = this._onNoClick.bind(this);
+        this._onYesClick = this._onYesClick.bind(this);
     }
 
     componentWillReceiveProps(nextProps) {
@@ -21,56 +27,124 @@ class Prompt extends Component {
     }
 
     render() {
-        const { children, className, inline, message, style } = this.props;
-        const { show } = this.state;
+        const { children, className, inline, inlineHorizontalAlign, inlineMessageColor, message, style } = this.props;
+        const { show, inlineVerticalAlign } = this.state;
         const containerClasses = ClassNames('ui', 'prompt', className, {
-            'promp-show': show,
+            'prompt-show': show,
             'prompt-inline': inline
         });
+        const messageClasses = ClassNames('prompt-message', {
+            'promp-message-alert': inlineMessageColor === 'alert' || children.props.color === 'alert' || children.props.buttonColor === 'alert',
+            'promp-message-success': inlineMessageColor === 'success' || children.props.color === 'success' || children.props.buttonColor === 'success'
+        });
+        const promptActionsStyle = {
+            left: !inlineHorizontalAlign || inlineHorizontalAlign === 'left' ? 0 : null,
+            right: inlineHorizontalAlign === 'right' ? 0 : null,
+            top: inlineVerticalAlign
+        };
 
         return (
-            <div clasName={containerClasses} onClick={this._onClick}>
+            <div className={containerClasses} style={style}>
                 {this._renderAction()}
 
-                <div>
-                    <div>{message || 'Are you sure?'}</div>
-                    <div>Yes</div>
-                    <div>No</div>
+                <div className="prompt-actions" style={promptActionsStyle}>
+                    <div className={messageClasses}>{message || 'Are you sure?'}</div>
+                    <a className="prompt-yes-btn" onClick={this._onYesClick}>{'Yes'}</a>
+                    <a className="prompt-no-btn" onClick={this._onNoClick}>{'No'}</a>
                 </div>
             </div>
         );
     }
 
-    _onClick(test) {
-        console.log('testing argument:', test);
+    componentDidMount() {
+        if (this.props.inline) {
+            this._findInlineVerticalPosition();
+        }
+    }
+
+    _findInlineVerticalPosition() {
+        const childHeight = ReactDOM.findDOMNode(this.promptAction).offsetHeight;
+        const negativeSpace = 5;
+        this.setState({ inlineVerticalAlign: childHeight + negativeSpace + 'px' });
+    }
+
+    _onClick(option) {
         const { onClick } = this.props;
         const { show } = this.state;
 
+        if (show) { return false; }
+
         if (!_.isUndefined(onClick)) {
-            onClick();
+            onClick(option);
         } else {
-            this.setState({ show: !show });
+            this.setState({ show: true });
+        }
+    }
+
+    _onNoClick() {
+        const { onNoClick } = this.props;
+
+        if (!_.isUndefined(onNoClick)) {
+            onNoClick();
+        } else {
+            this.setState({ show: false });
+        }
+    }
+
+    _onYesClick() {
+        const { onYesClick } = this.props;
+
+        if (!_.isUndefined(onYesClick)) {
+            onYesClick();
+        } else {
+            this.setState({ show: false });
         }
     }
 
     _renderAction() {
         const { children } = this.props;
+        const { show } = this.state;
+        const promptActionClasses = ClassNames('prompt-action', {
+            'prompt-action-disable': show
+        });
 
-        if (children) {
+        if (children && children.type.name === 'Button') {
             return React.cloneElement(children, {
-                className: 'promp-action',
-                onClick: this.onClickRef
+                className: promptActionClasses,
+                color: show ? 'disable' : children.props.color,
+                onClick: this._onClick.bind(this),
+                ref: promptAction => { this.promptAction = promptAction }
+            });
+        } else if (children && children.type.name === 'Dropdown') {
+            return React.cloneElement(children, {
+                className: promptActionClasses,
+                buttonColor: show ? 'disable' : children.props.buttonColor,
+                onChange: this._onClick.bind(this),
+                ref: promptAction => { this.promptAction = promptAction }
+            });
+        } else {
+            return React.cloneElement(children, {
+                className: `${promptActionClasses} ${show ? 'color-static' : null}`,
+                onClick: this._onClick.bind(this),
+                ref: promptAction => { this.promptAction = promptAction }
             });
         }
     }
 
 }
 
+const inlineHorizontalAlign = [ 'left', 'right' ];
+const inlineMessageColor = [ 'alert', 'success' ];
+
 Prompt.propTypes = {
     className: React.PropTypes.string,
     inline: React.PropTypes.bool,
+    inlineHorizontalAlign: React.PropTypes.oneOf(inlineHorizontalAlign),
+    inlineMessageColor: React.PropTypes.oneOf(inlineMessageColor),
     message: React.PropTypes.string,
     onClick: React.PropTypes.func,
+    onNoClick: React.PropTypes.func,
+    onYesClick: React.PropTypes.func,
     style: React.PropTypes.object
 };